Updated fluidflow
Lots of changes here. I changed a number of things to be recursive for convergence, instead of a set number of iterations, including Jesse's fancy grid stencil (Formerly Gauss-Seidel), and the contraint functions for periodic boundaries.

Gauss-Seidel within the projection step is unchanged, due to calculation hanging when introducing an error threshold.

Added a 0.99 multiplication to dampen the linear regression part of the advection. With it, the simulation works as expected. Without it, the simulation diverges. Unsure what the cause is, it needs to be investigated further.

Reworked the boundary conditions. They should more accurately reflect what occurs now.

Added in quiver representation of the velocity vector field. Note that it points in the opposite direction to the visualised fluid flow, because heatmap gonna heatmap.